ASoC: audio-graph: Fixup CPU endpoint hw_params in a BE<->BE link
When multiple components are connected back to back in an audio path,
hw_param fixup may be required for CPU or Codec endpoint of BE<->BE
DAI links. Currently fixup support is available for Codec and this
commit adds similar feature for CPU endpoint of a BE<->BE link.

For example a resampler component can be plugged into an audio path.
  [ FE -> BE1 -> ... -> resampler -> ... BEn ]

The resampler DAI links can be:
  BEx (CPU)              -> resampler input (Codec)
  resampler output (CPU) -> BEy (Codec)

Thus input and output sample rate parameters for resampler can be
fixed up as per the resample requirement.

    Signed-off-by: Sameer Pujar <spujar@nvidia.com>
    Cc: Kuninori Morimoto <kuninori.morimoto.gx@renesas.com>
    Link: https://lore.kernel.org/r/1631551342-25469-4-git-send-email-spujar@nvidia.com
    Signed-off-by: Mark Brown <broonie@kernel.org>

Author: Sameer Pujar <spujar@nvidia.com>
Date: Mon Sep 13 22:12:11 2021 +0530

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 30b428d02cbc9888d84407306d54dce8c2b8bfbf)
Bugzilla: https://bugzilla.redhat.com/2003613