ALSA: virtio: handling control and I/O messages for the PCM device
The driver implements a message-based transport for I/O substream
operations. Before the start of the substream, the hardware buffer is
sliced into I/O messages, the number of which is equal to the current
number of periods. The size of each message is equal to the current
size of one period.

I/O messages are organized in an ordered queue. The completion of the
I/O message indicates an elapsed period (the only exception is the end
of the stream for the capture substream). Upon completion, the message
is automatically re-added to the end of the queue.

    Signed-off-by: Anton Yakovlev <anton.yakovlev@opensynergy.com>
    Link: https://lore.kernel.org/r/20210302164709.3142702-6-anton.yakovlev@opensynergy.com
    Signed-off-by: Takashi Iwai <tiwai@suse.de>

Author: Anton Yakovlev <anton.yakovlev@opensynergy.com>
Date: Tue Mar 2 17:47:05 2021 +0100

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit f40a28679e0b7cb3a9cc6627a8dbb40961990f0a)
Bugzilla: https://bugzilla.redhat.com/2003613