Revert "ALSA: sb: fix a missing check of snd_ctl_add"
This reverts commit beae77170c60aa786f3e4599c18ead2854d8694d.

Because of recent interactions with developers from @umn.edu, all
commits from them have been recently re-reviewed to ensure if they were
correct or not.

Upon review, this commit was found to be incorrect for the reasons
below, so it must be reverted.  It is safe to ignore this error as the
mixer element is optional, and the driver is very legacy.

    Cc: Aditya Pakki <pakki001@umn.edu>
    Reviewed-by: Takashi Iwai <tiwai@suse.de>
    Link: https://lore.kernel.org/r/20210503115736.2104747-8-gregkh@linuxfoundation.org
    Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>

Author: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Date: Mon May 3 13:56:34 2021 +0200

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 4b059ce1f4b368208c2310925f49be77f15e527b)
Bugzilla: https://bugzilla.redhat.com/2003613