ALSA: es1688: Avoid devres management for es1688 object creation
The recent refactoring of es1688 object creation with the use of
devres caused a problem with the non-PnP probe of GUS driver, as it
tries to probe multiple times with different parameters  That is, this
object needs the immediate resource release and the devres doesn't fit
for it.

This patch reverts partially the commit for restoring the classic
resource management for es1688 object.

    Fixes: 1bb11c1c7f6e ("ALSA: es1688: Allocate resources with device-managed APIs")
    Reported-by: kernel test robot <oliver.sang@intel.com>
    Link: https://lore.kernel.org/r/20210805032513.GA30485@xsang-OptiPlex-9020
    Link: https://lore.kernel.org/r/20210805062148.30951-1-tiwai@suse.de
    Signed-off-by: Takashi Iwai <tiwai@suse.de>

Author: Takashi Iwai <tiwai@suse.de>
Date: Thu Aug 5 08:21:48 2021 +0200

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit df8bcf36be2755a39c2f1bcc1ea3c18329c29931)
Bugzilla: https://bugzilla.redhat.com/2003613