extcon: arizona: Disable mic detect if running when driver is removed
Microphone detection provides the button detection features on the
Arizona CODECs as such it will be running if the jack is currently
inserted. If the driver is unbound whilst the jack is still inserted
this will cause warnings from the regulator framework as the MICVDD
regulator is put but was never disabled.

Correct this by disabling microphone detection on driver removal and if
the microphone detection was running disable the regulator and put the
runtime reference that was currently held.

    Signed-off-by: Charles Keepax <ckeepax@opensource.cirrus.com>
    Signed-off-by: Chanwoo Choi <cw00.choi@samsung.com>

Author: Charles Keepax <ckeepax@opensource.cirrus.com>
Date: Thu Apr 4 17:33:56 2019 +0100

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 00053de52231117ddc154042549f2256183ffb86)
Bugzilla: https://bugzilla.redhat.com/2003613