ath11k: add support for WCN6855 hw2.1
Bugzilla: http://bugzilla.redhat.com/1970036
Bugzilla: http://bugzilla.redhat.com/1988269

Conflicts:
`struct ath11k_hw_params` had many changes between v5.15 and this patch, and
since we're applying it on top of v5.15 these changes are missing.

This patch adds a new entry to `ath11k_hw_params` array, which contains
`struct ath11k_hw_params` elements. Because of this I had to edit the changes
to this `ath11k_hw_params` array.

QCA confirmed that the new board being added to the array is identical to the
previous one except in it's name. Also this is the diff between the previous
board and the new one, made from untouched upstream commit d1147a316b:
```
  {
 -    .name = "wcn6855 hw2.0",
 -    .hw_rev = ATH11K_HW_WCN6855_HW20,
 +    .name = "wcn6855 hw2.1",
 +    .hw_rev = ATH11K_HW_WCN6855_HW21,
      .fw = {
 -        .dir = "WCN6855/hw2.0",
 +        .dir = "WCN6855/hw2.1",
          .board_size = 256 * 1024,
          .cal_offset = 128 * 1024,
      },
```

So, to resolve the conflict, I edited the changes to make that the new array
element is identical to the previous one but changing those 3 fields.

commit d1147a316b53df9cb0152e415ec41dcb6ea62c1c
Author: Baochen Qiang <quic_bqiang@quicinc.com>
Date:   Mon Nov 29 10:56:12 2021 +0800

    ath11k: add support for WCN6855 hw2.1

    Ath11k fails to probe WCN6855 hw2.1 chip:

    [ 6.983821] ath11k_pci 0000:06:00.0: enabling device (0000 -> 0002)
    [ 6.983841] ath11k_pci 0000:06:00.0: Unsupported WCN6855 SOC hardware version: 18 17

    This is caused by the wrong bit mask setting of hardware major version:
    for QCA6390/QCN6855, it should be BIT8-11, not BIT8-16, so change the
    definition to GENMASK(11, 8).

    Also, add a separate entry for WCN6855 hw2.1 in ath11k_hw_params.

    Please note that currently WCN6855 hw2.1 shares the same firmwares
    as hw2.0, so users of this chip need to create a symlink as below:

            ln -s hw2.0 hw2.1

    Tested-on: WCN6855 hw2.1 PCI WLAN.HSP.1.1-01720.1-QCAHSPSWPL_V1_V2_SILICONZ_LITE-1
    Tested-on: WCN6855 hw2.0 PCI WLAN.HSP.1.1-01720.1-QCAHSPSWPL_V1_V2_SILICONZ_LITE-1
    Tested-on: QCA6390 hw2.0 PCI WLAN.HST.1.0.1-01740-QCAHSTSWPLZ_V2_TO_X86-1

    Fixes: 18ac1665e785 ("ath11k: pci: check TCSR_SOC_HW_VERSION")
    Signed-off-by: Baochen Qiang <quic_bqiang@quicinc.com>
    Signed-off-by: Kalle Valo <quic_kvalo@quicinc.com>
    Link: https://lore.kernel.org/r/20211129025613.21594-1-quic_bqiang@quicinc.com

Signed-off-by: Íñigo Huguet <ihuguet@redhat.com>