revert ath11k: add support for WCN6855 hw2.1
Bugzilla: http://bugzilla.redhat.com/2043453  
Upstream-status: RHEL8 only

Commit d1147a316b53d (ath11k: add support for WCN6855 hw2.1 ) is a v5.17 commit,
but it was applied in RHEL 8 on top of the wireless update to v5.15 to bring
support to this hardware (RHEL hash 94f516d0016e2).

Between 5.15 and this commit, many changes were made to the same portion of
code. To avoid lot of conflicts in the following commits, let's revert it
and apply again later.

Signed-off-by: Íñigo Huguet <ihuguet@redhat.com>