Revert "ceph: enable async dirops by default"
Upstream Status: RHEL only
Bugzilla: https://bugzilla.redhat.com/2052542

This reverts commit d41418223872ae3dfe836492915fbab220975e9d.

Originally, I took this in because we had had it on upstream for a bit,
but we've had some bug reports recently that make me more leery of
changing the default in RHEL8.

Let's switch back to -o nowsync for now, and we'll shoot for making it
the default in RHEL9 and up.

Signed-off-by: Jeffrey Layton <jlayton@redhat.com>