ice: Fix kernel crash in XDP scenario
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=2069082
Upstream status: RHEL only

Commit 57f7f8b6bc0b ("ice: Use xdp_buf instead of rx_buf for xsk
zero-copy") introduced a regression that causes kernel crash
when an attempt to alloc a buffer from XSK pool fails. This
failure is not handled properly and dereference of NULL pointer
follows. Upstream is not affected because this code was refactored
by later commit that cannot be backported due to lack of batch
allocation in XDP core.

Signed-off-by: Ivan Vecera <ivecera@redhat.com>