cxl/acpi: Enumerate ports in RCD mode to enable RCHs and RCDs
Do the plumbing of ports to enable RCD/RCH pairs.

Do this by enumerating all necessary ports an endpoint needs to
connect to. This includes:

 1) A CXL root port with dport links to the RCHs. The port links to
    the CXL root platform device for RCH mode.

 2) RCH ports with dport links to its endpoints. Port connects to the
    pci host bridge device.

 3) CXL device endpoint connected to the RCH.

The port creation for the endpoint (3) is already implemented and
works in RCD mode too. Thus, it is not scope of this patch. Only the
endpoints must be registered at the host bridge port.

Implement this by introducing the function cxl_enumerate_rch_ports().
It registers a CXL host at the CXL root device, creates the host's
port and registers the existing CXL memory device endpoint at it. The
port of the CXL root device is created with the first CXL host being
registered.

Once enumerated, CXL restricted hosts show up in sysfs with CXL
devices connected as endpoints to it.

Signed-off-by: Robert Richter <rrichter@amd.com>